Move constants from polymesh-common-utilities to primitives.
