+++ pallets/contracts/src/benchmarking.rs
@@ -30,9 +30,9 @@ use wasm_instrument::parity_wasm::elements::{Instruction, ValueType};
 
 use pallet_identity::ParentDid;
 use polymesh_common_utilities::benchs::{cdd_provider, user, User, UserBuilder};
-use polymesh_common_utilities::constants::currency::POLY;
 use polymesh_common_utilities::group::GroupTrait;
 use polymesh_primitives::asset::AssetId;
+use polymesh_primitives::constants::currency::POLY;
 use polymesh_primitives::identity::limits::{
     MAX_ASSETS, MAX_EXTRINSICS, MAX_PALLETS, MAX_PORTFOLIOS,
 };
