+++ pallets/group/src/lib.rs
@@ -80,9 +80,9 @@ pub mod benchmarking;
 use pallet_identity as identity;
 pub use polymesh_common_utilities::{
     group::{Config, GroupTrait, InactiveMember, MemberCount, RawEvent, WeightInfo},
-    Context, GC_DID,
+    Context,
 };
-use polymesh_primitives::{committee::COMMITTEE_MEMBERS_MAX, IdentityId};
+use polymesh_primitives::{committee::COMMITTEE_MEMBERS_MAX, IdentityId, GC_DID};
 
 use frame_support::{
     decl_error, decl_module, decl_storage,
