+++ pallets/pips/src/lib.rs
@@ -108,16 +108,16 @@ use sp_version::RuntimeVersion;
 
 use pallet_base::{ensure_opt_string_limited, try_next_post};
 use pallet_identity::PermissionedCallOriginData;
-use polymesh_common_utilities::constants::PIP_MAX_REPORTING_SIZE;
 use polymesh_common_utilities::traits::balances::LockableCurrencyExt;
 use polymesh_common_utilities::traits::governance_group::GovernanceGroupTrait;
 use polymesh_common_utilities::traits::group::GroupTrait;
 use polymesh_common_utilities::traits::identity::Config as IdentityConfig;
-use polymesh_common_utilities::{with_transaction, CommonConfig, MaybeBlock};
-use polymesh_common_utilities::{GC_DID, TECHNICAL_DID, UPGRADE_DID};
+use polymesh_common_utilities::{CommonConfig, MaybeBlock};
+use polymesh_primitives::constants::PIP_MAX_REPORTING_SIZE;
 use polymesh_primitives::protocol_fee::{ChargeProtocolFee, ProtocolOp};
-use polymesh_primitives::storage_migration_ver;
+use polymesh_primitives::{storage_migration_ver, with_transaction};
 use polymesh_primitives::{Balance, IdentityId, Url};
+use polymesh_primitives::{GC_DID, TECHNICAL_DID, UPGRADE_DID};
 use polymesh_runtime_common::PipsEnactSnapshotMaximumWeight;
 
 use crate::types::{compare_spip, ProposalData, MAX_NORMAL_PRIORITY, PIPS_LOCK_ID};
