+++ pallets/runtime/develop/src/runtime.rs
@@ -26,8 +26,8 @@ use pallet_corporate_actions::ballot as pallet_corporate_ballot;
 use pallet_corporate_actions::distribution as pallet_capital_distribution;
 use pallet_session::historical as pallet_session_historical;
 pub use pallet_transaction_payment::{Multiplier, RuntimeDispatchInfo, TargetedFeeAdjustment};
-use polymesh_common_utilities::constants::currency::*;
-use polymesh_common_utilities::constants::ENSURED_MAX_LEN;
+use polymesh_primitives::constants::currency::*;
+use polymesh_primitives::constants::ENSURED_MAX_LEN;
 use polymesh_primitives::protocol_fee::ProtocolOp;
 use polymesh_primitives::settlement::Leg;
 use polymesh_primitives::{Balance, BlockNumber, Moment};
