+++ pallets/runtime/tests/src/asset_test.rs
@@ -23,7 +23,6 @@ use pallet_portfolio::{
 };
 use pallet_statistics::AssetStats;
 use polymesh_common_utilities::asset::AssetFnTrait;
-use polymesh_common_utilities::constants::currency::ONE_UNIT;
 use polymesh_common_utilities::traits::checkpoint::{
     NextCheckpoints, ScheduleCheckpoints, ScheduleId,
 };
@@ -37,6 +36,7 @@ use polymesh_primitives::asset_metadata::{
     AssetMetadataSpec, AssetMetadataValue, AssetMetadataValueDetail,
 };
 use polymesh_primitives::calendar::{CalendarPeriod, CalendarUnit, FixedOrVariableCalendarUnit};
+use polymesh_primitives::constants::currency::ONE_UNIT;
 use polymesh_primitives::settlement::{
     InstructionId, Leg, SettlementType, VenueDetails, VenueId, VenueType,
 };
