+++ pallets/settlement/src/benchmarking.rs
@@ -24,9 +24,9 @@ use sp_std::prelude::*;
 use pallet_asset::benchmarking::setup_asset_transfer;
 use pallet_nft::benchmarking::setup_nft_transfer;
 use polymesh_common_utilities::benchs::{create_and_issue_sample_asset, User, UserBuilder};
-use polymesh_common_utilities::constants::currency::ONE_UNIT;
-use polymesh_common_utilities::constants::ENSURED_MAX_LEN;
 use polymesh_primitives::checked_inc::CheckedInc;
+use polymesh_primitives::constants::currency::ONE_UNIT;
+use polymesh_primitives::constants::ENSURED_MAX_LEN;
 use polymesh_primitives::settlement::ReceiptMetadata;
 use polymesh_primitives::{IdentityId, Memo, NFTId, NFTs, PortfolioId, Ticker};
