+++ pallets/staking/src/pallet/impls.rs
@@ -58,8 +58,8 @@ use frame_support::traits::schedule::{DispatchTime, HIGHEST_PRIORITY};
 use frame_support::traits::DefensiveSaturating;
 
 use polymesh_common_utilities::identity::IdentityFnTrait;
-use polymesh_common_utilities::GC_DID;
 use polymesh_primitives::IdentityId;
+use polymesh_primitives::GC_DID;
 
 use crate::pallet::SlashingSwitch;
 use crate::UnlockChunk;
