+++ pallets/staking/src/pallet/mod.rs
@@ -56,9 +56,9 @@ use frame_system::offchain::SendTransactionTypes;
 use sp_runtime::traits::{AccountIdConversion, Dispatchable};
 use sp_runtime::Permill;
 
-use polymesh_common_utilities::constants::GC_PALLET_ID;
 use polymesh_common_utilities::identity::{Config as IdentityConfig, IdentityFnTrait};
-use polymesh_common_utilities::GC_DID;
+use polymesh_primitives::constants::GC_PALLET_ID;
+use polymesh_primitives::GC_DID;
 use polymesh_primitives::{storage_migration_ver, IdentityId};
 
 use crate::types::{PermissionedIdentityPrefs, SlashingSwitch};
