+++ pallets/sto/src/lib.rs
@@ -40,12 +40,12 @@ use pallet_identity::PermissionedCallOriginData;
 use pallet_settlement::VenueInfo;
 use polymesh_common_utilities::portfolio::PortfolioSubTrait;
 use polymesh_common_utilities::traits::{identity, portfolio};
-use polymesh_common_utilities::with_transaction;
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::impl_checked_inc;
 use polymesh_primitives::settlement::{Leg, ReceiptDetails, SettlementType, VenueId, VenueType};
 use polymesh_primitives::{
-    storage_migration_ver, Balance, EventDid, IdentityId, PortfolioId, WeightMeter,
+    storage_migration_ver, with_transaction, Balance, EventDid, IdentityId, PortfolioId,
+    WeightMeter,
 };
 use polymesh_primitives_derive::VecU8StrongTyped;
